+++ lib/src/state/actions/allStopsActions.dart
@@ -13,3 +13,5 @@ class AllStopsFetchFailure {
 
   AllStopsFetchFailure(this.allStopsErrorMessage);
 }
+
+class AllStopsClearError {}