+++ CHANGELOG.md
@@ -1,8 +1,17 @@
+# 1.1.1
+
+Minor bug fixes and improvements
+
+Showing a message if the search did not find any stops on explore screen
+Adjusting button in onboarding
+
 # 1.1.0
 
 HUGE improvements!!!
 
 Changed name from "Boston Subway Companion" to "meTro Boston"
+Added live vehicle locations on the map
+Added visual lines for each subway line on the map
 Updated app icon
 Improved search results on explore page
 Improved look and feel of search bars